pyproject.toml
[project]
name = "fiesta"
version = "0.0.0"
description = "Practise your Spanish by taking an AI teacher to an AI party."
authors = [
    {name = "Will Langdale", email = "willdon@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-anthropic>=0.3.0",
    "langchain-community>=0.3.8",
    "langgraph>=0.2.53",
    "python-dotenv>=1.0.1",
    "python-fasthtml>=0.10.0",
]

[dependency-groups]
dev = [
    "poethepoet>=0.31.0",
    "pytest>=8.3.3",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",
    "F",
    "I",
    "A",
    "B",
    "N",
    "W",
    "RUF",
]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.poe.tasks]
_check = "ruff check . --fix"
_format = "ruff format"

[tool.poe.tasks.run]
cmd = "uv run python src/fiesta/main.py"
help = "Run the FastHTML development server"

[tool.poe.tasks.format]
sequence = ["_check", "_format"]
help = "Run code formatting checks and fixes with ruff."
